ADD TEST for Service

+++ src/main/java/com/example/jeux_olimpique/service/UserserviceImpl.java
@@ -60,6 +60,10 @@ public class UserserviceImpl implements UserService {
 
 	@Override
 	public void deleteUser(Long id) {
+		
+		if(!userRepository.existsById(id)) {
+			throw new RuntimeException("user not find");
+		}
 		userRepository.deleteById(id);
 
 	}
